Fix virtual host config for Nginx proxy
Currently `vagrant up` results in an error when `proxy_enables` is set to `yes` and `proxy_ssl` is set to `no`.
This is caused by two mistakes. First of all the vhost-http template and vhost-https template are identical. As a result, the nginx tries to load `snippets/ssl` even if ssl is disabled, which then can not be found.
Second, the Apache2 must listen on port 80 as soon as the Nginx is enabled, which is set by `proxy_enabled` rather than `proxy_ssl`.